refactor: merged reset passwd create
