+++ internal/handler/user_handler.go
@@ -154,3 +154,34 @@ func (h *UserHandler) ResetPassword(c *fiber.Ctx) error {
 
 	return c.Status(fiber.StatusOK).JSON(res)
 }
+
+// ResendVerificationEmail godoc
+// @Summary Resend verification email
+// @Description Resend verification email
+// @Tags user
+// @Accept json
+// @Produce json
+// @Param user body dto.ResetPasswordCreateRequestBody true "token"
+// @Success 204 "email is sent to user successfully"
+// @Failure 400 {object} dto.ErrorResponse "your request is invalid
+// @Failure 500 {object} dto.ErrorResponse "system cannot resend verification email"
+// @Router /user/resetpassword [post]
+func (h *UserHandler) ResetPasswordCreate(c *fiber.Ctx) error {
+	body := new(dto.ResetPasswordCreateRequestBody)
+
+	if err := c.BodyParser(body); err != nil {
+		return apperror.BadRequestError(err, "your request is invalid")
+	}
+	validate := validator.New()
+
+	if err := validate.Struct(body); err != nil {
+		return apperror.BadRequestError(err, "your request body is incorrect")
+	}
+
+	err := h.userService.ResetPasswordCreate(c.Context(), body.Email)
+	if err != nil {
+		return err
+	}
+
+	return c.SendStatus(fiber.StatusNoContent)
+}
